Add a flag for the ownership of the encapsulated TObject in MonitorObject
By default the MO is the owner of the TObject. It is important for the processes that handle MOs by themselves,
e.g. Checker or clients or mergers. It is explicitely set to true (owns) when the objects is first published by
a task in order to leave the task (the user) the ownership of the object it created.
We could consider later to always transfer ownership to MO.